+++ test/package.json
@@ -1,6 +1,7 @@
 {
   "name": "myapp",
   "version": "0.1.0",
+  "private": "true",
   "bin": {
     "myapp": "bin/myapp.js"
   },
